+++ disco_aws_automation/disco_es.py
