+++ setup.py
@@ -78,7 +78,7 @@ setup(name='asiaq',
         "bin/disco_ssh.py",
         "bin/disco_chaos.py",
         "bin/disco_snapshot.py",
-        "bin/disco_elasticsearch.py",
+        "bin/disco_es.py",
         "jenkins/bake_common.sh",
         "jenkins/bake_all_phase1.sh",
         "jenkins/bake_all_phase2.sh",
